[prefactor] Parametrize kubeconfig creation logic to make it reusable
